+++ lib/api/contracts/active_symbols/active_symbols.dart
@@ -3,7 +3,8 @@ import 'package:flutter_deriv_api/basic_api/generated/api.dart';
 import 'package:flutter_deriv_api/api/models/active_symbols_model.dart';
 import 'package:flutter_deriv_api/services/connection/basic_binary_api.dart';
 import 'package:flutter_deriv_api/services/dependency_injector/injector.dart';
-import 'package:flutter_deriv_api/api/contracts/active_symbols/exceptions/active_symbols_exception.dart';
+
+import 'exceptions/active_symbols_exception.dart';
 
 /// Active Symbol
 class ActiveSymbol extends ActiveSymbolModel {
