+++ lib/api/contracts/contracts_for/contracts_for_symbol.dart
@@ -4,7 +4,8 @@ import 'package:flutter_deriv_api/api/models/contract_model.dart';
 import 'package:flutter_deriv_api/services/connection/basic_binary_api.dart';
 import 'package:flutter_deriv_api/api/models/contracts_for_symbol_model.dart';
 import 'package:flutter_deriv_api/services/dependency_injector/injector.dart';
-import 'package:flutter_deriv_api/api/contracts/contracts_for/exceptions/contract_for_symbol_exception.dart';
+
+import 'exceptions/contract_for_symbol_exception.dart';
 
 /// available contracts. Note: if the user is authenticated,
 /// then only contracts allowed under his account will be returned.
