+++ lib/api/models/contract_model.dart
@@ -1,5 +1,5 @@
-import 'package:flutter_deriv_api/api/models/base_model.dart';
-import 'package:flutter_deriv_api/api/models/forward_starting_option_model.dart';
+import 'base_model.dart';
+import 'forward_starting_option_model.dart';
 
 /// available contracts details
 class ContractModel extends BaseModel {
