+++ lib/api/models/contracts_for_symbol_model.dart
@@ -1,5 +1,5 @@
-import 'package:flutter_deriv_api/api/models/base_model.dart';
-import 'package:flutter_deriv_api/api/models/contract_model.dart';
+import 'base_model.dart';
+import 'contract_model.dart';
 
 /// Abstract class involving fields of [ContractsFor]
 abstract class ContractsForSymbolModel extends BaseModel {
